Rework workOrder system
Also restructure builder workOrder parts into a new abstract class.
With it abstract the workOrder systems from the job and building into abstract classes.
Create new workOrder types for each use case.
