+++ src/main/java/com/minecolonies/coremod/colony/buildings/AbstractBuildingStructureBuilder.java
@@ -1,0 +1,332 @@
+package com.minecolonies.coremod.colony.buildings;
+
+import com.minecolonies.api.util.CompatibilityUtils;
+import com.minecolonies.api.util.InventoryUtils;
+import com.minecolonies.api.util.ItemStackUtils;
+import com.minecolonies.coremod.colony.CitizenData;
+import com.minecolonies.coremod.colony.Colony;
+import com.minecolonies.coremod.colony.buildings.utils.BuildingBuilderResource;
+import com.minecolonies.coremod.colony.jobs.AbstractJobStructure;
+import com.minecolonies.coremod.colony.workorders.WorkOrderBuild;
+import com.minecolonies.coremod.colony.workorders.WorkOrderBuildDecoration;
+import com.minecolonies.coremod.inventory.InventoryCitizen;
+import io.netty.buffer.ByteBuf;
+import net.minecraft.item.ItemStack;
+import net.minecraft.nbt.NBTTagCompound;
+import net.minecraft.nbt.NBTTagList;
+import net.minecraft.tileentity.TileEntity;
+import net.minecraft.tileentity.TileEntityChest;
+import net.minecraft.util.EnumFacing;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.World;
+import net.minecraftforge.fml.common.network.ByteBufUtils;
+import net.minecraftforge.items.wrapper.InvWrapper;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+import java.util.HashMap;
+import java.util.Map;
+import java.util.function.Predicate;
+
+import static com.minecolonies.api.util.constant.ColonyConstants.NUM_ACHIEVEMENT_FIRST;
+import static com.minecolonies.api.util.constant.ToolLevelConstants.TOOL_LEVEL_WOOD_OR_GOLD;
+
+/**
+ * The structureBuilder building.
+ */
+public abstract class AbstractBuildingStructureBuilder extends AbstractBuildingWorker
+{
+    /**
+     * The maximum upgrade of the building.
+     */
+    public static final  int    MAX_BUILDING_LEVEL = 5;
+
+    /**
+     * Tags to store the needed resourced to nbt.
+     */
+    private static final String TAG_RESOURCE_LIST = "resourcesItem";
+
+    /**
+     * Contains all resources needed for a certain build.
+     */
+    private HashMap<String, BuildingBuilderResource> neededResources = new HashMap<>();
+
+    /**
+     * Public constructor of the building, creates an object of the building.
+     *
+     * @param c the colony.
+     * @param l the position.
+     */
+    public AbstractBuildingStructureBuilder(final Colony c, final BlockPos l)
+    {
+        super(c, l);
+    }
+
+    /**
+     * Getter of the max building level.
+     *
+     * @return the integer.
+     */
+    @Override
+    public int getMaxBuildingLevel()
+    {
+        return MAX_BUILDING_LEVEL;
+    }
+
+    @Override
+    public Map<Predicate<ItemStack>, Integer> getRequiredItemsAndAmount()
+    {
+        final Map<Predicate<ItemStack>, Integer> toKeep = new HashMap<>(keepX);
+        toKeep.putAll(keepX);
+
+        for (final BuildingBuilderResource stack : neededResources.values())
+        {
+            toKeep.put(itemstack -> ItemStackUtils.compareItemStacksIgnoreStackSize(stack.getItemStack(), itemstack, true, true), stack.getAmount());
+        }
+
+        return toKeep;
+    }
+
+    @Override
+    public ItemStack forceTransferStack(final ItemStack stack, final World world)
+    {
+        final ItemStack itemStack = super.forceTransferStack(stack, world);
+        if (ItemStackUtils.isEmpty(itemStack))
+        {
+            this.markDirty();
+        }
+
+        return itemStack;
+    }
+
+    @Override
+    public void readFromNBT(@NotNull final NBTTagCompound compound)
+    {
+        super.readFromNBT(compound);
+        final NBTTagList neededResTagList = compound.getTagList(TAG_RESOURCE_LIST, net.minecraftforge.common.util.Constants.NBT.TAG_COMPOUND);
+        for (int i = 0; i < neededResTagList.tagCount(); ++i)
+        {
+            final NBTTagCompound neededRes = neededResTagList.getCompoundTagAt(i);
+            final ItemStack stack = new ItemStack(neededRes);
+            final BuildingBuilderResource resource = new BuildingBuilderResource(stack, ItemStackUtils.getSize(stack));
+            final int hashCode = stack.hasTagCompound() ? stack.getTagCompound().hashCode() : 0;
+            neededResources.put(stack.getUnlocalizedName() + ":" + stack.getItemDamage() + "-" + hashCode, resource);
+        }
+    }
+
+    @Override
+    public void writeToNBT(@NotNull final NBTTagCompound compound)
+    {
+        super.writeToNBT(compound);
+        @NotNull final NBTTagList neededResTagList = new NBTTagList();
+        for (@NotNull final BuildingBuilderResource resource : neededResources.values())
+        {
+            @NotNull final NBTTagCompound neededRes = new NBTTagCompound();
+            final ItemStack itemStack = new ItemStack(resource.getItem(), resource.getAmount(), resource.getDamageValue());
+            itemStack.setTagCompound(resource.getItemStack().getTagCompound());
+            itemStack.writeToNBT(neededRes);
+
+            neededResTagList.appendTag(neededRes);
+        }
+        compound.setTag(TAG_RESOURCE_LIST, neededResTagList);
+    }
+
+    /**
+     * Method to serialize data to send it to the view.
+     *
+     * @param buf the used ByteBuffer.
+     */
+    @Override
+    public void serializeToView(@NotNull final ByteBuf buf)
+    {
+        super.serializeToView(buf);
+
+        updateAvailableResources();
+        buf.writeInt(neededResources.size());
+        for (@NotNull final BuildingBuilderResource resource : neededResources.values())
+        {
+            ByteBufUtils.writeItemStack(buf, resource.getItemStack());
+            buf.writeInt(resource.getAvailable());
+            buf.writeInt(resource.getAmount());
+        }
+
+        final CitizenData data = this.getMainWorker();
+        if(data != null && data.getJob() instanceof AbstractJobStructure)
+        {
+            final AbstractJobStructure structureBuilderJob = (AbstractJobStructure) data.getJob();
+            final WorkOrderBuildDecoration workOrderBuildDecoration = structureBuilderJob.getWorkOrder();
+            if(workOrderBuildDecoration != null)
+            {
+                final BlockPos pos = workOrderBuildDecoration.getBuildingLocation();
+                final String name =
+                        workOrderBuildDecoration instanceof WorkOrderBuild ? ((WorkOrderBuild) workOrderBuildDecoration).getUpgradeName() : workOrderBuildDecoration.getName();
+                ByteBufUtils.writeUTF8String(buf, name);
+
+                final String desc;
+                if(pos.equals(getLocation()))
+                {
+                    desc = "here";
+                }
+                else
+                {
+                    final BlockPos relativePos = getLocation().subtract(pos);
+                    final EnumFacing facingX = EnumFacing.getFacingFromVector(relativePos.getX(), 0, 0);
+                    final EnumFacing facingZ = EnumFacing.getFacingFromVector(0, 0, relativePos.getZ());
+                    desc = relativePos.getX() + " " + facingX + " " + relativePos.getZ() + " " + facingZ;
+                }
+
+                ByteBufUtils.writeUTF8String(buf, desc);
+            }
+            else
+            {
+                ByteBufUtils.writeUTF8String(buf, "-");
+                ByteBufUtils.writeUTF8String(buf, "");
+            }
+        }
+        else
+        {
+            ByteBufUtils.writeUTF8String(buf, "-");
+            ByteBufUtils.writeUTF8String(buf, "");
+        }
+    }
+
+    /**
+     * Update the available resources.
+     * <p>
+     * which are needed for the build and in the structureBuilder's chest or inventory
+     */
+    private void updateAvailableResources()
+    {
+        getMainWorkerEntity().ifPresent(structureBuilder -> {
+            final InventoryCitizen structureBuilderInventory = getMainWorker().getInventory();
+            if (structureBuilderInventory == null)
+            {
+                return;
+            }
+
+            for (@NotNull final Map.Entry<String, BuildingBuilderResource> entry : neededResources.entrySet())
+            {
+                final BuildingBuilderResource resource = entry.getValue();
+
+                resource.setAvailable(0);
+
+                if (structureBuilderInventory != null)
+                {
+
+                    resource.addAvailable(InventoryUtils.getItemCountInItemHandler(new InvWrapper(structureBuilderInventory),
+                            stack -> ItemStackUtils.compareItemStacksIgnoreStackSize(stack, resource.getItemStack(), true, true)));
+                }
+
+                final TileEntity chestInventory = this.getTileEntity();
+                if (chestInventory != null)
+                {
+                    resource.addAvailable(InventoryUtils.getItemCountInProvider(chestInventory,
+                            stack -> ItemStackUtils.compareItemStacksIgnoreStackSize(stack, resource.getItemStack(), true, true)));
+                }
+
+                //Count in the additional chests as well
+                if (structureBuilder != null)
+                {
+                    for (final BlockPos pos : getAdditionalCountainers())
+                    {
+                        final TileEntity entity = CompatibilityUtils.getWorld(structureBuilder).getTileEntity(pos);
+                        if (entity instanceof TileEntityChest)
+                        {
+                            resource.addAvailable(InventoryUtils.getItemCountInProvider(entity,
+                                    stack -> ItemStackUtils.compareItemStacksIgnoreStackSize(stack, resource.getItemStack(), true, true)));
+                        }
+                    }
+                }
+            }
+        });
+    }
+
+    /**
+     * Get the needed resources for the current build.
+     *
+     * @return a new Hashmap.
+     */
+    public Map<String, BuildingBuilderResource> getNeededResources()
+    {
+        return new HashMap<>(neededResources);
+    }
+
+    /**
+     * Add a new resource to the needed list.
+     *
+     * @param res    the resource.
+     * @param amount the amount.
+     */
+    public void addNeededResource(@Nullable final ItemStack res, final int amount)
+    {
+        if (ItemStackUtils.isEmpty(res) || amount == 0)
+        {
+            return;
+        }
+        final int hashCode = res.hasTagCompound() ? res.getTagCompound().hashCode() : 0;
+        BuildingBuilderResource resource = this.neededResources.get(res.getUnlocalizedName() + ":" + res.getItemDamage() + "-" + hashCode);
+        if (resource == null)
+        {
+            resource = new BuildingBuilderResource(res, amount);
+        }
+        else
+        {
+            resource.setAmount(resource.getAmount() + amount);
+        }
+        this.neededResources.put(res.getUnlocalizedName() + ":" + res.getItemDamage() + "-" + hashCode, resource);
+        this.markDirty();
+    }
+
+    /**
+     * Reduce a resource of the needed list.
+     *
+     * @param res    the resource.
+     * @param amount the amount.
+     */
+    public void reduceNeededResource(final ItemStack res, final int amount)
+    {
+        final int hashCode = res.hasTagCompound() ? res.getTagCompound().hashCode() : 0;
+        int preAmount = 0;
+        final String name = res.getUnlocalizedName() + ":" + res.getItemDamage() + "-" + hashCode;
+        if (this.neededResources.containsKey(name))
+        {
+            preAmount = this.neededResources.get(name).getAmount();
+        }
+
+        if (preAmount - amount <= 0)
+        {
+            this.neededResources.remove(name);
+        }
+        else
+        {
+            this.neededResources.get(name).setAmount(preAmount - amount);
+        }
+        this.markDirty();
+    }
+
+    /**
+     * Resets the needed resources completely.
+     */
+    public void resetNeededResources()
+    {
+        neededResources = new HashMap<>();
+        this.markDirty();
+    }
+
+    /**
+     * Check if the structureBuilder requires a certain ItemStack for the current construction.
+     *
+     * @param stack the stack to test.
+     * @return true if so.
+     */
+    public boolean requiresResourceForBuilding(final ItemStack stack)
+    {
+        final int hashCode = stack.hasTagCompound() ? stack.getTagCompound().hashCode() : 0;
+        return neededResources.containsKey(stack.getUnlocalizedName() + ":" + stack.getItemDamage() + "-" + hashCode);
+    }
+
+    /**
+     * Search a workOrder for the worker.
+     */
+    public abstract void searchWorkOrder();
+}
