+++ src/main/java/com/minecolonies/coremod/colony/jobs/AbstractJobStructure.java
@@ -1,13 +1,29 @@
 package com.minecolonies.coremod.colony.jobs;
 
 import com.minecolonies.coremod.colony.CitizenData;
+import com.minecolonies.coremod.colony.buildings.AbstractBuilding;
+import com.minecolonies.coremod.colony.buildings.BuildingBuilder;
+import com.minecolonies.coremod.colony.workorders.WorkOrderBuildDecoration;
 import com.minecolonies.coremod.util.StructureWrapper;
+import net.minecraft.nbt.NBTTagCompound;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 /**
  * Common job object for all structure AIs.
  */
 public abstract class AbstractJobStructure extends AbstractJob
 {
+    /**
+     * Tag to store the workOrder id.
+     */
+    private static final String TAG_WORK_ORDER = "workorder";
+
+    /**
+     * The id of the current workOrder.
+     */
+    private int workOrderId;
+
     /**
      * The structure the job should build.
      */
@@ -53,5 +69,98 @@ public abstract class AbstractJobStructure extends AbstractJob
     public void setStructure(final StructureWrapper structure)
     {
         this.structure = structure;
+    }
+
+    /**
+     * Get the Work Order ID for this Job.
+     *
+     * @return UUID of the Work Order claimed by this Job, or null
+     */
+    public int getWorkOrderId()
+    {
+        return workOrderId;
+    }
+
+    /**
+     * Does this job have a Work Order it has claimed?
+     *
+     * @return true if there is a Work Order claimed by this Job
+     */
+    public boolean hasWorkOrder()
+    {
+        return workOrderId != 0;
+    }
+
+    @Override
+    public void readFromNBT(@NotNull final NBTTagCompound compound)
+    {
+        super.readFromNBT(compound);
+        if (compound.hasKey(TAG_WORK_ORDER))
+        {
+            workOrderId = compound.getInteger(TAG_WORK_ORDER);
+        }
+    }
+
+    @Override
+    public void writeToNBT(@NotNull final NBTTagCompound compound)
+    {
+        super.writeToNBT(compound);
+        if (workOrderId != 0)
+        {
+            compound.setInteger(TAG_WORK_ORDER, workOrderId);
+        }
+    }
+
+    /**
+     * Do final completion when the Job's current work is complete.
+     */
+    public void complete()
+    {
+        getWorkOrder().onCompleted(getCitizen().getColony());
+        getCitizen().getColony().getWorkManager().removeWorkOrder(workOrderId);
+        setWorkOrder(null);
+        setStructure(null);
+        this.getColony().getStatsManager().incrementStatistic("huts");
+    }
+
+    /**
+     * Get the Work Order for the Job.
+     * Warning: WorkOrder is not cached
+     *
+     * @return WorkOrderBuildDecoration for the Build
+     */
+    public WorkOrderBuildDecoration getWorkOrder()
+    {
+        return getColony().getWorkManager().getWorkOrder(workOrderId, WorkOrderBuildDecoration.class);
+    }
+
+    /**
+     * Reset the needed items list.
+     */
+    private void resetNeededItems()
+    {
+        final AbstractBuilding workerBuilding = this.getCitizen().getWorkBuilding();
+        if (workerBuilding instanceof BuildingBuilder)
+        {
+            ((BuildingBuilder) workerBuilding).resetNeededResources();
+        }
+    }
+
+    /**
+     * Set a Work Order for this Job.
+     *
+     * @param order Work Order to associate with this job, or null
+     */
+    public void setWorkOrder(@Nullable final WorkOrderBuildDecoration order)
+    {
+        if (order == null)
+        {
+            workOrderId = 0;
+            resetNeededItems();
+        }
+        else
+        {
+            workOrderId = order.getID();
+        }
     }
 }
